Complete refactor of build system
