+++ .gitignore
@@ -1,1 +1,5 @@
 **/.venv/
+
+# Ignore the dockerfile directory (generated during build)
+dockerfiles/
+__pycache__/
