+++ layers/numpy.json
@@ -8,12 +8,21 @@
         "RUN pip freeze --path . > requirements.txt"
     ],
     "runtimes": {
+        "python3.10": {
+            "default_image": "public.ecr.aws/lambda/python:3.10",
+            "default_layer_target_directory": "/python/lib/python3.10/site-packages",
+            "versions": {
+                "2.2.6": {
+                    "$ref": "#/$defs/version-2.2.6"
+                }
+            }
+        },
         "python3.11": {
             "default_image": "public.ecr.aws/lambda/python:3.11",
             "default_layer_target_directory": "/python/lib/python3.11/site-packages",
             "versions": {
-                "2.4.0": {
-                    "$ref": "#/$defs/version-2.4.0"
+                "2.2.6": {
+                    "$ref": "#/$defs/version-2.2.6"
                 }
             }
         },
@@ -21,8 +30,8 @@
             "default_image": "public.ecr.aws/lambda/python:3.12",
             "default_layer_target_directory": "/python/lib/python3.12/site-packages",
             "versions": {
-                "2.4.0": {
-                    "$ref": "#/$defs/version-2.4.0"
+                "2.4.1": {
+                    "$ref": "#/$defs/version-2.4.1"
                 }
             }
         },
@@ -30,8 +39,8 @@
             "default_image": "public.ecr.aws/lambda/python:3.13",
             "default_layer_target_directory": "/python/lib/python3.13/site-packages",
             "versions": {
-                "2.4.0": {
-                    "$ref": "#/$defs/version-2.4.0"
+                "2.4.1": {
+                    "$ref": "#/$defs/version-2.4.1"
                 }
             }
         },
@@ -39,16 +48,25 @@
             "default_image": "public.ecr.aws/lambda/python:3.14",
             "default_layer_target_directory": "/python/lib/python3.14/site-packages",
             "versions": {
-                "2.4.0": {
-                    "$ref": "#/$defs/version-2.4.0"
+                "2.4.1": {
+                    "$ref": "#/$defs/version-2.4.1"
                 }
             }
         }
     },
     "$defs": {
-        "version-2.4.0": {
+        "version-2.2.6": {
             "common_instructions_pre": [
-                "RUN pip install numpy==2.4.0 --target ."
+                "RUN pip install numpy==2.2.6 --target ."
+            ],
+            "architectures": {
+                "x86_64": {},
+                "arm64": {}
+            }
+        },
+        "version-2.4.1": {
+            "common_instructions_pre": [
+                "RUN pip install numpy==2.4.1 --target ."
             ],
             "architectures": {
                 "x86_64": {},
